backend/requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
faiss-cpu
pandas
python-multipart
numpy
scikit-learn
